feat: refactor useSubscriptionError
